requirements.txt
streamlit
chromadb
sentence-transformers
requests
python-dotenv
PyPDF2
python-docx 